tests/e2e/requirements.txt
aiobotocore==2.21.0
aiofiles==24.1.0
aiogram==3.18.0
aiogram_dialog==2.3.1
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aioitertools==0.12.0
aiosignal==1.3.2
aiosqlite==0.21.0
alabaster==0.7.16
allure-pytest==2.13.5
allure-python-commons==2.13.5
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
asyncpg==0.30.0
attrs==25.1.0
babel==2.17.0
blinker==1.9.0
botocore==1.37.1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
colorlog==6.9.0
commonmark==0.9.1
coverage==7.6.10
cryptography==44.0.1
distlib==0.3.9
docopt==0.6.2
docutils==0.21.2
exceptiongroup==1.2.2
execnet==2.1.1
Faker==35.2.0
fastapi==0.115.8
filelock==3.17.0
Flask==3.1.0
flit==3.10.1
flit_core==3.10.1
fluent-logger==0.11.1
frozenlist==1.5.0
google-api-core==2.24.1
google-api-python-client==2.160.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpc-interceptor==0.15.4
grpcio==1.70.0
grpcio-reflection==1.62.3
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
identify==2.6.6
idna==3.10
imagesize==1.4.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
magic-filter==1.0.12
Markdown==3.7
MarkupSafe==3.0.2
msgpack==1.1.0
multidict==6.1.0
nodeenv==1.9.1
packaging==24.2
paho-mqtt==1.6.1
passlib==1.7.4
pbr==6.1.1
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.1.0
propcache==0.3.0
proto-plus==1.26.0
protobuf==4.25.6
py==1.11.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
pykwalify==1.8.0
pyparsing==3.2.1
pyproject-api==1.9.0
pytest==7.2.2
pytest-cov==6.0.0
pytest-xdist==3.6.1
python-box==6.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
recommonmark==0.7.1
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
rsa==4.9
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
ruff==0.9.4
setuptools==75.8.2
six==1.17.0
sniffio==1.3.1
snowballstemmer==2.2.0
Sphinx==7.4.7
sphinx-markdown-tables==0.0.17
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqladmin==0.20.1
SQLAlchemy==2.0.38
starlette==0.45.3
stevedore==4.1.1
tavern==2.11.0
tomli==2.2.1
tomli_w==1.2.0
tox==4.24.1
types-aiobotocore-s3==2.20.0
types-jmespath==1.0.2.20240106
types-jsonschema==4.23.0.20241208
types-paho-mqtt==1.6.0.20240321
types-protobuf==4.25.0.20240417
types-PyYAML==6.0.12.20241230
types-requests==2.32.0.20241016
typing_extensions==4.12.2
tzlocal==5.3
uritemplate==4.1.1
urllib3==2.3.0
uv==0.5.26
uvicorn==0.34.0
virtualenv==20.29.1
watchfiles==1.0.4
websockets==15.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
WTForms==3.1.2
yarl==1.18.3
